scripts/requirements.txt
# mongodb
pymongo

# cassandra
cassandra-driver